initramfs-live-install/efi: set grub timeout to 0
The final grub menu that is created after installation
only gets a single boot entry and it is also set as
default along with a timeout. The usability of the menu
is minimal in such a case so we now set the timeout
entry to 0 to make grub follow on to booting as soon
as possible.

Signed-off-by: Awais Belal <awais_belal@mentor.com>