core-image-sato: add libglu, glew and glut in default image
These libraries actually get pulled in through the dependency
from mesa-demos but they are still required by the customer
when 500MB RFS limit is enforced in which case we actually
remove mesa-demos from the RFS and hence the mentioned
libraries aren't deployed either.
This commit also rearranges the extra package addition
mechanism to make it more clear and widely usable.

Signed-off-by: Awais Belal <awais_belal@mentor.com>