pachinko-sdk-ogl: resolve deployment issues
These changes allow deployment of pachinko-sdk-ogl
at the same time with the pachinko-sdk package. One
provides support for the Vulkan based SDK and the
other targets OpenGL. The modifications are made
for the OpenGL version and the user visible impact
will be
1: use tiledPlayback-OGL if one wants to use the
OGL version of the demo.
2: explicitly provide
LD_LIRARY_PATH=/usr/lib64/pachinko-sdk-ogl
when running tiledPlayback-OGL demo.

Signed-off-by: Awais Belal <awais_belal@mentor.com>